[#IC-105] split test files
